+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to
 - @cosmjs/stargate: Handle key value pairs in tx search correctly if the value
   is a numeric value. ([#1462])
 
-[#1462] : https://github.com/cosmos/cosmjs/issues/1462
+[#1462]: https://github.com/cosmos/cosmjs/issues/1462
 
 ### Changed